Added weights and biases initialisation logic in the neural network

+++ src/neural_network/layer_trait.rs
@@ -10,11 +10,11 @@ pub trait Layer<T> {
 
     fn get_output_shape(&self) -> &Vec<usize>;
 
-    fn get_weights(&self) -> &Array<i16, Ix1>;
+    fn get_weights(&self) -> &Array<i32, Ix1>;
 
-    fn get_bias(&self) -> i16;
+    fn get_bias(&self) -> i32;
 
     // In the process of training the model, the weights and the bias will be changed multiple times
-    fn change_weights(&mut self, new_weights:Array<i16, Ix1>);
-    fn change_bias(&mut self, new_bias:i16);
+    fn change_weights(&mut self, new_weights:Array<i32, Ix1>);
+    fn change_bias(&mut self, new_bias:i32);
 }
